// apps/super-app/src/utils/commons/enums.ts
export enum ESTATUS_WORKING {
  ACTIVE = "ACTIVE",
  INACTIVE = "INACTIVE",
}

export enum ECURRENCY {
  USD = "CURRENCY_USD",
  EUR = "CURRENCY_EUR",
  GBP = "CURRENCY_GBP",
  JPY = "CURRENCY_JPY",
  AUD = "CURRENCY_AUD",
  CAD = "CURRENCY_CAD",
  CHF = "CURRENCY_CHF",
  HKD = "CURRENCY_HKD",
  SGD = "CURRENCY_SGD",
  SEK = "CURRENCY_SEK",
  ARS = "CURRENCY_ARS",
  BRL = "CURRENCY_BRL",
  CNY = "CURRENCY_CNY",
  COP = "CURRENCY_COP",
  CZK = "CURRENCY_CZK",
  DKK = "CURRENCY_DKK",
  HUF = "CURRENCY_HUF",
  ILS = "CURRENCY_ILS",
  INR = "CURRENCY_INR",
  KRW = "CURRENCY_KRW",
  MXN = "CURRENCY_MXN",
  NOK = "CURRENCY_NOK",
  NZD = "CURRENCY_NZD",
  PLN = "CURRENCY_PLN",
  RUB = "CURRENCY_RUB",
  THB = "CURRENCY_THB",
  TRY = "CURRENCY_TRY",
  TWD = "CURRENCY_TWD",
  UAH = "CURRENCY_UAH",
  VND = "CURRENCY_VND",
  ZAR = "CURRENCY_ZAR",
}

export enum EDURATION_UNIT {
  YEAR = "DURATION_UNIT_YEAR",
  WEEK = "DURATION_UNIT_WEEK",
  DAY = "DURATION_UNIT_DAY",
  MONTH = "DURATION_UNIT_MONTH",
  QUARTERLY = "DURATION_UNIT_QUARTER",
}

export enum EPAYMENT_VENDOR {
  STRIPE = "PAYMENT_PROVIDER_STRIPE",
}
export enum EPAYMENT_METHOD {
  CARD = "PAYMENT_METHOD_CARD",
  BANK_TRANSFER = "PAYMENT_METHOD_BANK_TRANSFER",
  WALLET = "PAYMENT_METHOD_WALLET",
}

export enum EUSER_ROLE {
  PREMIUM = "USER_ROLE_PREMIUM",
  FREE = "USER_ROLE_FREE",
  CANCELLED_PREMIUM = "USER_ROLE_CANCELLED_PREMIUM",
}

export enum EPRODUCT_STATUS {
  ACTIVE = "SUBSCRIPTION_PACKAGE_STATUS_ACTIVE",
  INACTIVE = "SUBSCRIPTION_PACKAGE_STATUS_INACTIVE",
}

export enum ESUBSCRIPTION_SOURCE {
  WEB = "SUBSCRIPTION_SOURCE_ECOSYSTEM",
  ANDROID = "SUBSCRIPTION_SOURCE_INAPP_ANDROID",
  IOS = "SUBSCRIPTION_SOURCE_INAPP_IOS",
  STRIPE = "SUBSCRIPTION_SOURCE_STRIPE",
}

export enum ESUBSCRIPTION_STATUS {
  ACTIVE = "SUBSCRIPTION_STATUS_ACTIVE", // Subscription is active
  INACTIVE = "SUBSCRIPTION_STATUS_INACTIVE", // Subscription is inactive but can be reactivated
  CANCELLED = "SUBSCRIPTION_STATUS_CANCELLED", // Subscription has been cancelled
  EXPIRED = "SUBSCRIPTION_STATUS_EXPIRED", // Subscription has expired
  TRIAL = "SUBSCRIPTION_STATUS_TRIAL", // Subscription is in trial period
  PENDING = "SUBSCRIPTION_STATUS_PENDING", // Subscription is pending (e.g., awaiting payment),
  SUSPENDED = "SUBSCRIPTION_STATUS_SUSPENDED", // Subscription is suspended due to non-payment or other issues,
  GRACE_PERIOD = "SUBSCRIPTION_STATUS_GRACE_PERIOD_STARTED", // Grace period has started after expiration or payment failure
}

export enum EAUTH_PROVIDER {
  GOOGLE = "google",
  APPLE = "apple",
  FACEBOOK = "facebook",
}

export enum EAUTH_SOURCE {
  DEFAULT = "Default",
  GOOGLE_ONE_TAP = "google_onetap",
}

export enum EPAYMENT_VENDOR_NUMBER {
  PAYMENT_PROVIDER_UNSPECIFIED = 0, // (nếu user mua gói ở mobile)
  PAYMENT_PROVIDER_STRIPE = 1,
  PAYMENT_PROVIDER_PADDLE = 2,
}